requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-tenants==3.6.0
django-cors-headers==4.5.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.2.0
django-redis==5.4.0
django-environ==0.11.2
django-extensions==3.2.3
whitenoise==6.8.2
Pillow==10.4.0
python-dateutil==2.9.0

